[MOVE/#25] 카테고리버튼 컴포넌트 코드 위치 변경
