+++ app/src/main/java/com/sopt/shinmungo/presentation/report/component/CategoryGrid.kt
@@ -6,11 +6,14 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.GridItemSpan
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
+import com.sopt.shinmungo.core.designsystem.component.button.RoundedCornerTextButton
+import com.sopt.shinmungo.core.designsystem.theme.ShinMunGoTheme
 import com.sopt.shinmungo.presentation.report.ReportViewModel
 
 @Composable
@@ -60,4 +63,24 @@ fun CategoryGrid(viewModel: ReportViewModel, modifier: Modifier = Modifier) {
             }
         }
     }
+}
+
+@Composable
+fun CategoryButton(
+    category: String,
+    isClicked: Boolean,
+    onCategoryButtonClick: (String) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    RoundedCornerTextButton(
+        text = category,
+        textStyle = ShinMunGoTheme.typography.body6,
+        textColor = if (isClicked) ShinMunGoTheme.color.gray1 else ShinMunGoTheme.color.gray13,
+        backgroundColor = if (isClicked) ShinMunGoTheme.color.primary else ShinMunGoTheme.color.gray1,
+        roundedCornerShape = RoundedCornerShape(10.dp),
+        onButtonClick = {
+            onCategoryButtonClick(category)
+        },
+        modifier = Modifier.padding(vertical = 16.dp)
+    )
 }
